requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
email-validator>=2.0.0
langgraph>=0.2.0,<1.0.0
langchain-groq>=0.1.0,<1.0.0
tavily-python>=0.5.0,<1.0.0
aiohttp>=3.9.0,<4.0.0
google-auth-oauthlib>=1.3.0
google-auth-httplib2>=0.3.0
google-api-python-client>=2.100.0

